+++ tw-gaffer-jta-starter/src/main/java/com/transferwise/common/gaffer/starter/GafferJtaConfiguration.java
@@ -1,9 +1,19 @@
 package com.transferwise.common.gaffer.starter;
 
-import com.transferwise.common.gaffer.ServiceRegistry;
-import com.transferwise.common.gaffer.ServiceRegistryHolder;
+import com.transferwise.common.baseutils.meters.cache.IMeterCache;
+import com.transferwise.common.baseutils.meters.cache.MeterCache;
+import com.transferwise.common.gaffer.DefaultGafferTransactionManager;
+import com.transferwise.common.gaffer.DefaultMetricsTemplate;
+import com.transferwise.common.gaffer.GafferJtaProperties;
+import com.transferwise.common.gaffer.GafferTransactionManager;
+import com.transferwise.common.gaffer.GafferUserTransaction;
+import com.transferwise.common.gaffer.MetricsTemplate;
+import com.transferwise.common.gaffer.util.Clock;
+import com.transferwise.common.gaffer.util.MonotonicClock;
+import io.micrometer.core.instrument.MeterRegistry;
 import jakarta.transaction.TransactionManager;
 import jakarta.transaction.UserTransaction;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -16,26 +26,42 @@ import org.springframework.transaction.jta.JtaTransactionManager;
 public class GafferJtaConfiguration {
 
   @Bean
-  @ConditionalOnMissingBean
-  public UserTransaction gafferJtaUserTransaction() {
-    ServiceRegistry serviceRegistry = ServiceRegistryHolder.getServiceRegistry();
-    return serviceRegistry.getUserTransaction();
+  @ConditionalOnMissingBean(IMeterCache.class)
+  public IMeterCache twDefaultMeterCache(MeterRegistry meterRegistry) {
+    return new MeterCache(meterRegistry);
   }
 
   @Bean
-  @ConditionalOnMissingBean
-  public TransactionManager gafferJtaTransactionManager(GafferJtaProperties properties) {
-    ServiceRegistry serviceRegistry = ServiceRegistryHolder.getServiceRegistry();
-    serviceRegistry.getConfiguration().setBeforeCommitValidationRequiredTimeMs(properties.getBeforeCommitValidationRequiredTime().toMillis());
-    return serviceRegistry.getTransactionManager();
+  @ConditionalOnMissingBean(Clock.class)
+  public MonotonicClock gafferJtaClock() {
+    return new MonotonicClock();
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(MetricsTemplate.class)
+  public DefaultMetricsTemplate gafferJtaMetricsTemplate(IMeterCache meterCache) {
+    return new DefaultMetricsTemplate(meterCache);
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(TransactionManager.class)
+  public DefaultGafferTransactionManager gafferJtaTransactionManager(GafferJtaProperties gafferJtaProperties, MetricsTemplate metricsTemplate,
+      Clock clock) {
+    return new DefaultGafferTransactionManager(gafferJtaProperties, metricsTemplate, clock);
+  }
+
+  @Bean
+  @ConditionalOnBean(GafferTransactionManager.class)
+  @ConditionalOnMissingBean(UserTransaction.class)
+  public UserTransaction gafferJtaUserTransaction(GafferTransactionManager gafferTransactionManager) {
+    return new GafferUserTransaction(gafferTransactionManager);
   }
 
   @Bean("transactionManager")
   @ConditionalOnMissingBean
-  public JtaTransactionManager gafferJtaJtaTransactionManager(UserTransaction userTransaction, TransactionManager transactionManager) {
-    ServiceRegistry serviceRegistry = ServiceRegistryHolder.getServiceRegistry();
+  public JtaTransactionManager gafferJtaJtaTransactionManager(UserTransaction userTransaction, GafferTransactionManager transactionManager) {
     JtaTransactionManager jtaTransactionManager = new JtaTransactionManager(userTransaction, transactionManager);
-    jtaTransactionManager.setTransactionSynchronizationRegistry(serviceRegistry.getTransactionSynchronizationRegistry());
+    jtaTransactionManager.setTransactionSynchronizationRegistry(transactionManager.getTransactionSynchronizationRegistry());
     jtaTransactionManager.setAllowCustomIsolationLevels(true);
     return jtaTransactionManager;
   }