+++ tw-gaffer-jta-starter/src/main/java/com/transferwise/common/gaffer/starter/GafferJtaDataSourceBeanProcessor.java
@@ -2,6 +2,8 @@ package com.transferwise.common.gaffer.starter;
 
 import com.transferwise.common.baseutils.ExceptionUtils;
 import com.transferwise.common.baseutils.jdbc.DataSourceProxyUtils;
+import com.transferwise.common.gaffer.GafferJtaProperties;
+import com.transferwise.common.gaffer.GafferTransactionManager;
 import com.transferwise.common.gaffer.jdbc.GafferJtaDataSource;
 import com.zaxxer.hikari.HikariDataSource;
 import javax.sql.DataSource;
@@ -20,16 +22,14 @@ public class GafferJtaDataSourceBeanProcessor implements BeanPostProcessor, Orde
   @Override
   public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
     return ExceptionUtils.doUnchecked(() -> {
-      if (!(bean instanceof DataSource)) {
+      if (!(bean instanceof DataSource dataSource)) {
         return bean;
       }
 
-      var dataSource = (DataSource) bean;
-
       if (dataSource.isWrapperFor(GafferJtaDataSource.class)) {
         // Why add the starter, if service already wrapped it by itself?
-        log.warn("Datasource '" + dataSource
-            + "' is already wrapped with `DataSourceImpl`. Remove the custom wrapping or `tw-gaffer-jta-starter` dependency.");
+        log.warn("Datasource '{}'"
+            + " is already wrapped with `DataSourceImpl`. Remove the custom wrapping or `tw-gaffer-jta-starter` dependency.", dataSource);
         return dataSource;
       }
 
@@ -44,16 +44,13 @@ public class GafferJtaDataSourceBeanProcessor implements BeanPostProcessor, Orde
       }
 
       final var properties = beanFactory.getBean(GafferJtaProperties.class);
-      final var databaseProperties = properties.getDatabases().get(databaseName);
+      final var transactionManager = beanFactory.getBean(GafferTransactionManager.class);
 
-      var gafferJtaDataSource = new GafferJtaDataSource(dataSource);
-      DataSourceProxyUtils.tieTogether(gafferJtaDataSource, dataSource);
-      gafferJtaDataSource.setUniqueName(databaseName);
-      gafferJtaDataSource.setRegisterAsMBean(false);
+      final var databaseProperties = properties.getDatabases().get(databaseName);
+      var gafferJtaDataSource = new GafferJtaDataSource(transactionManager, databaseName, dataSource);
 
       if (databaseProperties != null) {
-        gafferJtaDataSource.setRegisterAsMBean(databaseProperties.isRegisterAsMbean());
-        gafferJtaDataSource.setOrder(databaseProperties.getCommitOrder());
+        gafferJtaDataSource.setCommitOrder(databaseProperties.getCommitOrder());
         gafferJtaDataSource.setBeforeReleaseAutoCommitStrategy(databaseProperties.getAutoCommitStrategy());
         gafferJtaDataSource.setValidationTimeoutSeconds((int) databaseProperties.getConnectionValidationInterval().toSeconds());
       }