+++ tw-gaffer-jta/src/main/java/com/transferwise/common/gaffer/util/Uid.java
@@ -4,8 +4,6 @@ public interface Uid {
 
   long getStartTimeMillis();
 
-  int getSequence();
-
   String getInstanceId();
 
   byte[] asBytes();