+++ tw-gaffer-jta/src/test/java/com/transferwise/common/gaffer/test/complextest1/app/Client.java
@@ -4,12 +4,14 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import java.io.Serial;
 import java.io.Serializable;
 
 @Entity
 @Table(name = "clients")
 public class Client implements Serializable {
 
+  @Serial
   private static final long serialVersionUID = 1L;
   private Long id;
   private String name;